refactor(frontend): share raw-app source→draft-value projection across chat and deploy page
